DEBT-0016 residual closed (natStringCharAt uint16 leak -> int32 + TestNativeReturnWidths boundary contract); DEBT-0017 closed via StackEffect unification, nanoTime(J) AOT path verified (JProbe pin); docs catch-up: status.md M3 close, emitter-abi v0.2 as-built + stack backfill section, debt register active=5, P-0003..0006 archived, P-0007/0008 retro stubs, ADR-0010 representation, P-0009 M4 plan opened

+++ internal/kernel/natives.go
@@ -74,7 +74,11 @@ func natStringCharAt(ctx *CallContext, recv Value, args []Value) (Value, error) 
 		return nil, ctx.Throw("java/lang/StringIndexOutOfBoundsException",
 			"String index out of range: "+FormatInt(i))
 	}
-	return chars[i], nil
+	// Width discipline (DEBT-0016 residual): JVM values cross the
+	// native/Java boundary as canonical int32. Returning raw uint16 here
+	// survived the interpreter (frame.push normalizes) but panicked the
+	// AOT path's .(int32) assertions.
+	return int32(chars[i]), nil
 }
 
 func natStringEquals(ctx *CallContext, recv Value, args []Value) (Value, error) {
